requirements.txt
python-telegram-bot>=20.0
mistralai
python-dotenv
notion-client
rich
pydantic
transformers
httpx
# Optional utilities
aiohttp
typing-extensions
fastapi
uvicorn
pngrok  # for tunnel testing
